Created filters folder and added select team members filter to it
